basic group apis
