+++ packages/auth/src/hashing.js
@@ -1,5 +1,6 @@
 const bcrypt = require("bcryptjs")
 const env = require("./environment")
+const { v4 } = require("uuid")
 
 const SALT_ROUNDS = env.SALT_ROUNDS || 10
 
@@ -10,4 +11,8 @@ exports.hash = async data => {
 
 exports.compare = async (data, encrypted) => {
   return bcrypt.compare(data, encrypted)
+}
+
+exports.newid = function() {
+  return v4().replace(/-/g, "")
 }
