+++ packages/worker/src/api/routes/index.js
@@ -1,5 +1,5 @@
-const adminRoutes = require("./admin")
+const { userRoutes, groupRoutes } = require("./admin")
 const authRoutes = require("./auth")
 const appRoutes = require("./app")
 
-exports.routes = [adminRoutes, authRoutes, appRoutes]
+exports.routes = [userRoutes, groupRoutes, authRoutes, appRoutes]
