+++ packages/worker/src/constants/index.js
@@ -2,3 +2,7 @@ exports.UserStatus = {
   ACTIVE: "active",
   INACTIVE: "inactive",
 }
+
+exports.Groups = {
+  ALL_USERS: "all_users",
+}
